Revert changes on Muons and apply it on BaseCkfTrajectoryBuilder
