+++ RecoTracker/CkfPattern/interface/CkfTrajectoryBuilder.h
@@ -79,9 +79,9 @@ protected:
   unsigned int limitedCandidates(const TrajectorySeed& seed,
                                  TempTrajectory& startingTraj,
                                  TrajectoryContainer& result) const;
-  virtual unsigned int limitedCandidates(const std::shared_ptr<const TrajectorySeed>& sharedSeed,
-                                         TempTrajectoryContainer& candidates,
-                                         TrajectoryContainer& result) const;
+  unsigned int limitedCandidates(const std::shared_ptr<const TrajectorySeed>& sharedSeed,
+                                 TempTrajectoryContainer& candidates,
+                                 TrajectoryContainer& result) const;
 
   void updateTrajectory(TempTrajectory& traj, TM&& tm) const;
